Added camera rotation and relative movement
Added camera rotation to rotate around the player, this caused movement to break and movement now needed to take rotation of the camera (and the modified position due to this), into account.
My selected solution to this problem is to track the player's old and current position, get the difference and then apply that to the camera. So far this is working without issue.

The next issue caused by this was that the player was still moving based on the world's forward and right directions instead of the camera's forward and right direction.
This caused the movement to become difficult to predict and so the way to fix this was to apply the movement based on the camera's transform.forward and transform.right, this is currently working without issue.

Started working on a way of updating the player's position when they're on a moving platform so as to not slowly slip off or in the case of the elevator falling, suddenly drop.
My thoughts were to do a similar thing to how I solved the camera movement after implementing rotation but this seems to not work. I imagine because the player is always getting both positions from the platform as the exact same position due to the time taken for the fixed update to complete.
Need to look into either delaying the update of the old position based on a small amount of time or another solution to this issue.